Rearranged test case example loading code
