+++ src/MEDS_transforms/stages/aggregate_code_metadata/with_not_split_defined_shards/out_metadata.yaml
