+++ src/MEDS_transforms/stages/aggregate_code_metadata/with_split_dependent_shards/out_metadata.yaml
