+++ src/MEDS_transforms/stages/fit_vocabulary_indices/out_metadata.yaml
